+++ internal/events/channel_close.go
@@ -24,6 +24,8 @@ type ChannelCloseTemplate struct {
 	ChannelPoint    string
 	RemotePubkey    string
 	Capacity        string
+	CloseInitiator  bool
+	CloseType       int32
 }
 
 func NewChannelCloseEvent(node *lnrpc.LightningNode, channel *lnrpc.ChannelCloseSummary) *ChannelCloseEvent {
@@ -52,6 +54,8 @@ func (e *ChannelCloseEvent) GetTemplateData(lang language.Tag) interface{} {
 		Capacity:        format.FormatBasic(float64(e.Channel.Capacity), lang),
 		RemotePubkey:    e.Channel.RemotePubkey,
 		SettledBalance:  format.FormatBasic(float64(e.Channel.SettledBalance), lang),
+		CloseInitiator:  e.Channel.CloseInitiator == lnrpc.Initiator_INITIATOR_LOCAL,
+		CloseType:       int32(e.Channel.CloseType),
 	}
 }