+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 
+- New event: opening channel notification (@Feelancer21)
+- New event: closing channel notification (@Feelancer21)
 - Configuration option for number formatting locale (e.g., "en-US" for English, "de-DE" for German) (@Primexz)
+- New template variables `{{.CloseInitiator}}` and `{{.CloseType}}` for `channel_close_event`. (@Feelancer21)
 
 ### Fixed
 ### Changed